adjusted seed file

+++ server/db/models/order.js
@@ -3,7 +3,6 @@ const db = require('../db')
 
 const Order = db.define('orders', {
   //order date: automatically included as createdAt field
-
   stillInCart: {
     type: Sequelize.BOOLEAN,
     allowNull: false,
@@ -17,8 +16,7 @@ const Order = db.define('orders', {
     }
   },
   pricePaid: {
-    type: Sequelize.DECIMAL(10, 2)
-    //want to use integer & work in pennies & do calculation in a hook (to save as money?)
+    type: Sequelize.INTEGER
   }
 })
